A2Aserver/requirements.txt
langchain
langchain-community
langchain-google-genai
python-dotenv
fastapi
uvicorn
python-a2a
a2a-sdk
a2a-sdk[http-server]
a2a-sdk[all]
httpx